Renamed some files and reorganized to subdirectories
