+++ app/components/layout/Layout.tsx
@@ -10,7 +10,7 @@ import log from 'loglevel';
 
 import Home from '../../containers/Home';
 import Header from '../../containers/layout/Header';
-import PageManagement from '../templates/PageManagement';
+import PageManagement from '../templateManagement/pages/PageManagement';
 import Consts from '../../utils/consts';
 import { NO_PAGES_FOR_TENANT_ERR, UNAUTHORIZED_ERR } from '../../utils/ErrorCodes';
 import SplashLoadingScreen from '../../utils/SplashLoadingScreen';
@@ -18,7 +18,7 @@ import SplashLoadingScreen from '../../utils/SplashLoadingScreen';
 import StatusPoller from '../../utils/StatusPoller';
 import UserAppDataAutoSaver from '../../utils/UserAppDataAutoSaver';
 import ScrollToTop from './ScrollToTop';
-import TemplateManagement from '../templates/TemplateManagement';
+import TemplateManagement from '../templateManagement/TemplateManagement';
 
 export default class Layout extends Component {
     static initialState = {
