+++ app/components/templateManagement/TemplateManagement.tsx
@@ -10,8 +10,8 @@ import { useDispatch, useSelector } from 'react-redux';
 import { push } from 'connected-react-router';
 import Const from '../../utils/consts';
 import { Breadcrumb, Button, Divider, ErrorMessage, Segment } from '../basic';
-import Pages from './Pages';
-import Templates from './Templates';
+import Pages from './pages/Pages';
+import Templates from './templates/Templates';
 import { createPageId, setActive } from '../../actions/templateManagement';
 import { selectHomePage } from '../../actions/pageMenu';
 import Internal from '../../utils/Internal';
