+++ app/components/templateManagement/pages/CreatePageModal.tsx
@@ -6,8 +6,8 @@ import _ from 'lodash';
 import PropTypes from 'prop-types';
 import React, { useState } from 'react';
 import i18n from 'i18next';
-import { useBoolean, useErrors, useResettableState } from '../../utils/hooks';
-import { ApproveButton, Button, CancelButton, Form, Icon, Modal } from '../basic/index';
+import { useBoolean, useErrors, useResettableState } from '../../../utils/hooks';
+import { ApproveButton, Button, CancelButton, Form, Icon, Modal } from '../../basic';
 
 export default function CreatePageModal({ onCreatePage, pageName: initialPageName }) {
     const [pageName, setPageName] = useState(initialPageName);
