+++ app/components/templateManagement/pages/PageManagement.tsx
@@ -9,16 +9,16 @@ import { arrayMove } from 'react-sortable-hoc';
 import type { ThunkDispatch } from 'redux-thunk';
 import type { AnyAction } from 'redux';
 
-import { Alert, Breadcrumb, Button, Divider, EditableLabel, ErrorMessage, Menu, Segment, Sidebar } from '../basic';
-import EditModeBubble from '../EditModeBubble';
-import { PageContent } from '../shared/widgets';
-import { createPageId, drillDownWarning, savePage, setActive, setPageEditMode } from '../../actions/templateManagement';
-import StageUtils from '../../utils/stageUtils';
-import { useErrors } from '../../utils/hooks';
-import { forEachWidget, getWidgetDefinitionById, LayoutSection, SimpleWidgetObj, TabContent } from '../../actions/page';
-import type { ReduxState } from '../../reducers';
-import type { WidgetDefinition } from '../../utils/StageAPI';
-import type { TemplatePageDefinition } from '../../reducers/templatesReducer';
+import { Alert, Breadcrumb, Button, Divider, EditableLabel, ErrorMessage, Menu, Segment, Sidebar } from '../../basic';
+import EditModeBubble from '../../EditModeBubble';
+import { PageContent } from '../../shared/widgets';
+import { createPageId, drillDownWarning, savePage, setActive, setPageEditMode } from '../../../actions/templateManagement';
+import StageUtils from '../../../utils/stageUtils';
+import { useErrors } from '../../../utils/hooks';
+import { forEachWidget, getWidgetDefinitionById, LayoutSection, SimpleWidgetObj, TabContent } from '../../../actions/page';
+import type { ReduxState } from '../../../reducers';
+import type { WidgetDefinition } from '../../../utils/StageAPI';
+import type { TemplatePageDefinition } from '../../../reducers/templatesReducer';
 
 export interface PageManagementProps {
     pageId: string;
