+++ app/components/templateManagement/templates/TenantList.tsx
@@ -7,20 +7,21 @@ import _ from 'lodash';
 import PropTypes from 'prop-types';
 import React from 'react';
 import i18n from 'i18next';
-import { Segment, Icon, Divider, List, Message, PopupConfirm } from '../basic';
+import Const from '../../../utils/consts';
+import { Segment, Icon, Divider, List, Message, PopupConfirm } from '../../basic';
 
-export default function RoleList({ custom, onDelete, roles, style }) {
+export default function TenantList({ custom, onDelete, style, tenants }) {
     return (
         <Segment style={style}>
-            <Icon name="student" /> Roles
+            <Icon name="male" /> Tenants
             <Divider />
             <List divided relaxed verticalAlign="middle" className="light">
-                {roles.map(item => {
+                {tenants.map(item => {
                     return (
                         <List.Item key={item}>
-                            {item}
+                            {item === Const.DEFAULT_ALL ? 'all' : item}
 
-                            {custom && _.size(roles) > 1 && (
+                            {custom && _.size(tenants) > 1 && (
                                 <PopupConfirm
                                     trigger={
                                         <Icon
@@ -31,8 +32,8 @@ export default function RoleList({ custom, onDelete, roles, style }) {
                                         />
                                     }
                                     content={i18n.t(
-                                        'templates.templateManagement.roleList.removeConfirm',
-                                        'Are you sure to remove this role from template?'
+                                        'templates.templateManagement.tenantsList.removeConfirm',
+                                        'Are you sure to remove this tenant from template?'
                                     )}
                                     onConfirm={() => onDelete(item)}
                                 />
@@ -40,24 +41,26 @@ export default function RoleList({ custom, onDelete, roles, style }) {
                         </List.Item>
                     );
                 })}
-                {_.isEmpty(roles) && (
-                    <Message content={i18n.t('templates.templateManagement.roleList.noRoles', 'No roles available')} />
+                {_.isEmpty(tenants) && (
+                    <Message
+                        content={i18n.t('templates.templateManagement.tenantsList.noTenants', 'No tenants available')}
+                    />
                 )}
             </List>
         </Segment>
     );
 }
 
-RoleList.propTypes = {
+TenantList.propTypes = {
     custom: PropTypes.bool,
     onDelete: PropTypes.func,
-    roles: PropTypes.arrayOf(PropTypes.string),
-    style: PropTypes.shape({})
+    style: PropTypes.shape({}),
+    tenants: PropTypes.arrayOf(PropTypes.string)
 };
 
-RoleList.defaultProps = {
+TenantList.defaultProps = {
     custom: false,
     onDelete: _.noop,
-    roles: [],
-    style: {}
+    style: {},
+    tenants: []
 };
